+++ packages/chart/src/Chart.tsx
@@ -19,9 +19,17 @@ import {
   DateTimeFormatSettings,
 } from '@deephaven/jsapi-utils';
 import Log from '@deephaven/log';
-import { Layout, Icon, PlotData } from 'plotly.js';
+import {
+  Config as PlotlyConfig,
+  Layout,
+  Icon,
+  Data,
+  PlotData,
+  ModeBarButtonAny,
+} from 'plotly.js';
+import type { PlotParams } from 'react-plotly.js';
+import createPlotlyComponent from 'react-plotly.js/factory.js';
 import Plotly from './plotly/Plotly';
-import Plot from './plotly/Plot';
 import ChartModel from './ChartModel';
 import ChartUtils, { ChartModelSettings } from './ChartUtils';
 import './Chart.scss';
@@ -38,6 +46,7 @@ interface ChartProps {
   model: ChartModel;
   settings: FormatterSettings;
   isActive: boolean;
+  Plotly: typeof Plotly;
   onDisconnect: () => void;
   onReconnect: () => void;
   onUpdate: (obj: { isLoading: boolean }) => void;
@@ -46,7 +55,7 @@ interface ChartProps {
 }
 
 interface ChartState {
-  data: Partial<PlotData>[] | null;
+  data: Partial<Data>[] | null;
   downsamplingError: unknown;
   isDownsampleFinished: boolean;
   isDownsampleInProgress: boolean;
@@ -65,6 +74,7 @@ export class Chart extends Component<ChartProps, ChartState> {
       showTSeparator: true,
       formatter: [],
     },
+    Plotly,
     onDisconnect: (): void => undefined,
     onReconnect: (): void => undefined,
     onUpdate: (): void => undefined,
@@ -124,6 +134,7 @@ export class Chart extends Component<ChartProps, ChartState> {
     this.handleRelayout = this.handleRelayout.bind(this);
     this.handleRestyle = this.handleRestyle.bind(this);
 
+    this.PlotComponent = createPlotlyComponent(props.Plotly);
     this.plot = React.createRef();
     this.plotWrapper = React.createRef();
     this.columnFormats = [];
@@ -180,7 +191,9 @@ export class Chart extends Component<ChartProps, ChartState> {
 
   currentSeries: number;
 
-  plot: RefObject<typeof Plot>;
+  PlotComponent: React.ComponentType<PlotParams>;
+
+  plot: RefObject<typeof this.PlotComponent>;
 
   plotWrapper: RefObject<HTMLDivElement>;
 
@@ -205,13 +218,15 @@ export class Chart extends Component<ChartProps, ChartState> {
       downsamplingError: unknown,
       isDownsampleFinished: boolean,
       isDownsampleInProgress: boolean,
-      isDownsamplingDisabled: boolean
-    ) => {
-      const customButtons = [];
+      isDownsamplingDisabled: boolean,
+      data: Partial<Data>[]
+    ): Partial<PlotlyConfig> => {
+      const customButtons: ModeBarButtonAny[] = [];
       const hasDownsampleError = Boolean(downsamplingError);
       if (hasDownsampleError) {
         customButtons.push({
           name: `Downsampling failed: ${downsamplingError}`,
+          title: 'Downsampling failed',
           click: () => undefined,
           icon: Chart.convertIcon(dhWarningFilled),
           attr: 'fill-warning',
@@ -236,11 +251,31 @@ export class Chart extends Component<ChartProps, ChartState> {
         const icon = isDownsampleInProgress ? vsLoading : dhGraphLineDown;
         customButtons.push({
           name,
+          title: 'Downsampling status',
           icon: Chart.convertIcon(icon),
           click: this.handleDownsampleClick,
           attr,
         });
       }
+
+      const has2D = data.some(
+        ({ type }) => type != null && !type.includes('3d')
+      );
+      const has3D = data.some(
+        ({ type }) => type != null && type.includes('3d')
+      );
+
+      const buttons2D = [
+        'zoomIn2d',
+        'zoomOut2d',
+        'autoScale2d',
+        'resetScale2d',
+      ] as const;
+      const buttons3D = [
+        'orbitRotation',
+        'tableRotation',
+        'resetCameraDefault3d',
+      ] as const;
 
       return {
         displaylogo: false,
@@ -249,14 +284,16 @@ export class Chart extends Component<ChartProps, ChartState> {
         // Yes, the value is a boolean or the string 'hover': https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js#L249
         displayModeBar:
           // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
-          isDownsampleInProgress || hasDownsampleError ? true : 'hover',
+          isDownsampleInProgress || hasDownsampleError
+            ? true
+            : ('hover' as const),
 
         // Each array gets grouped together in the mode bar
         modeBarButtons: [
           customButtons,
           ['toImage'],
-          ['zoom2d', 'pan2d'],
-          ['zoomIn2d', 'zoomOut2d', 'autoScale2d', 'resetScale2d'],
+          ['zoom2d', 'pan2d'], // These work the same for both 2d and 3d
+          [...(has2D ? buttons2D : []), ...(has3D ? buttons3D : [])],
         ],
       };
     }
@@ -398,7 +435,7 @@ export class Chart extends Component<ChartProps, ChartState> {
     }
   }
 
-  handlePlotUpdate(figure: { layout: Layout }): void {
+  handlePlotUpdate(figure: Readonly<{ layout: Partial<Layout> }>): void {
     // User could have modified zoom/pan here, update the model dimensions
     // We don't need to update the datarevision, as we don't have any data changes
     // until an update comes back from the server anyway
@@ -427,7 +464,7 @@ export class Chart extends Component<ChartProps, ChartState> {
     this.updateModelDimensions();
   }
 
-  handleRestyle([changes, seriesIndexes]: [
+  handleRestyle([changes, seriesIndexes]: readonly [
     Record<string, unknown>,
     number[]
   ]): void {
@@ -436,7 +473,7 @@ export class Chart extends Component<ChartProps, ChartState> {
       const { data } = this.state;
       const { onSettingsChanged } = this.props;
       if (data != null) {
-        const hiddenSeries = data.reduce(
+        const hiddenSeries = (data as Partial<PlotData>[]).reduce(
           (acc: string[], { name, visible }) =>
             name != null && visible === 'legendonly' ? [...acc, name] : acc,
           []
@@ -518,6 +555,7 @@ export class Chart extends Component<ChartProps, ChartState> {
 
   updateDimensions(): void {
     const rect = this.getPlotRect();
+    const { Plotly: PlotlyProp } = this.props;
     if (
       this.plot.current != null &&
       rect != null &&
@@ -526,15 +564,19 @@ export class Chart extends Component<ChartProps, ChartState> {
     ) {
       // Call relayout to resize avoiding the debouncing plotly does
       // https://github.com/plotly/plotly.js/issues/2769#issuecomment-402099552
-      Plotly.relayout(this.plot.current.el, { autosize: true }).catch(
-        (e: unknown) => {
-          log.debug('Unable to resize, promise rejected', e);
+      PlotlyProp.relayout(
+        ((this.plot.current as unknown) as { el: HTMLElement }).el,
+        {
+          autosize: true,
         }
-      );
+      ).catch((e: unknown) => {
+        log.debug('Unable to resize, promise rejected', e);
+      });
     }
   }
 
   render(): ReactElement {
+    const { PlotComponent } = this;
     const {
       data,
       downsamplingError,
@@ -548,13 +590,16 @@ export class Chart extends Component<ChartProps, ChartState> {
       downsamplingError,
       isDownsampleFinished,
       isDownsampleInProgress,
-      isDownsamplingDisabled
+      isDownsamplingDisabled,
+      data ?? []
     );
     const isPlotShown = data != null;
     return (
       <div className="h-100 w-100 chart-wrapper" ref={this.plotWrapper}>
         {isPlotShown && (
-          <Plot
+          <PlotComponent
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
             ref={this.plot}
             data={data}
             layout={layout}